graphs folder with graph components
